Fix bug where dependencies weren't being detected
The only issue is that now shortpaths has to be restructured in order to address
the other major issues. The updates for the shortpaths should be stored
separately instead of being in the done in the same iter() as the
shortpaths are being expanded/folded. Additionally, another enum for the
shortpath dependency should be made.